fix duplicate LeafletViews
React 18 wants us to have one static 'root' for every element that we are going to render into.
Unfortunately with Ionic's collection-repeat, DOM elements are reused and we can't guarantee that the same DOM element should be used for the same content. We'll have to reassign the root every time there's a change in the Angular element. Otherwise we get duplicates.
This will, unfortunately, clutter our console with warnings from React -- but it is only temporary until inf scroll list is converted to React.