create TimestampBadge component
In the label screen, the timestamps for each trip/place/untracked are shown at the top and bottom of the card. Since this is used in many places, it has been refactored into a reusable React component.
It uses the Badge component from react-native-paper, and applies the appropriate colors from the theme.