replace multilabel-ui with MultiLabelButtonGroup
MultiLabelButtonGroup.jsx will be the React component that handles the multilabel buttons. This is a React rewrite of multi-label-ui.
The popup menu is a work in progress; it uses a `Menu` from React Native Paper, but it does not render correctly because of the existing DOM structure that is from AngularJS. We need a way for this to appear on top of all other content.