finish MultiLabelButtonGroup functionality
At this point, the multilabel buttons are now behaving almost exactly the same as they used to - except rendered in React instead of AngularJS.

The popups are going to stay as ionic for now. For that to work, we use the $ionicPopover service from within React. This works fine for now.
Later, I'd prefer to make these into Menus or Dialogs. But I can't do this until we have some of the upper-level parts of the UI rewritten in React.