enable esModuleInterop
This allows us to 'import React from "react"', instead of having to 'import * as React from "react"'.
In this project we want to be consistent about using ES import syntax, preferably treating everything as either a default or named export. esModuleInterop allows us to do this even if the libraries differ.