add UserInputButton.jsx
This new component replaces enketo-trip-button. It has been made more generic; verbage has been changed to 'timelineEntry' in a few places that it still said 'trip'.
The enketoTripButton directive has been removed.
As was the case with the other buttons, EnketoTripButtonService is still in use and the enketo-trip-button.js file must be kept for now.