create AddedNotesList.tsx
This component replaces enketo-notes-list.
Under the hood, this is using DataTable from React Native Paper (but styled for our purposes). It does not look significantly different than the old implementation, and it should behave exactly the same.
The popup to confirm a deletion is still using $ionicPopup and will be revisited later.

Removing enketo-notes-list.js, notes_list.html, and styles relating to notes-list, since these are all now irrelevant.