add appTheme.ts
This extracts the theme setup into a separate file, and also allows us to define theme 'flavors', which are variations of the theme that can be used in different parts of the app.
This should allow us a more flexible, dynamic way of theming, in which we can use different color palettes in different parts of the app.
And in the future if we want to change a color, we'll just refer back to one place it is defined in appTheme.ts. This should also make it easy to implement a dark theme later on.