add DiaryCard.tsx
In the label view, we'll show trips, untracked time, and/or places. These will all have their own separate component.
DiaryCard supercedes all of these; any of the commonality between all cards will be contained here in DiaryCard.
This includes the RN Paper Card component, the logic to apply a 'flavor' of the theme to the card, and the timestamp badges at the top & bottom of each card.