extract filter select and date input to components
These are the filter select (ie. To Label / All Trips) and the date input used on the Label screen in the top bar.
For the Label screen itself to be converted, these are the last 'widgets' that need to be rewritten.

But, I am not sure what approach we want to take with these; React Native Paper doesn't have built-in dropdown menus, nor a datepicker (although there are some solid options created by the open source community).

These are going to stay partially as HTML for now, while I migrate the Label screen and we can decide later what to move forward with.