add TimelineScrollList
This component will replace infinite_scroll_list; it encompasses the ENTIRE label screen, including the header.
I'm sure that this leaves a lot of things broken, but it's a necessary step that had to be taken on.

At the core of the new list view is FlashList, a more performant implementation of React Native's FlatList. I spent some time looking for a performant scroll view that would work with both web and native; I think this is our best option.
FlashList is better optimized for native, but it seems to be performing as well or better than our previous Ionic implementation. Once we have gone fully native, we should see a BIG performance improvement thanks to this.

infinite_scroll_list.js was a very messy file. I think that the conversion helped clean it up some, but there are definitely more optimizations that can be done. A lot of this logic should be extracted elsewhere, and I'd like to revisit at a later time.