fix conflicting header bars of Ionic and RN Paper
Now that the whole label screen has been rewritten, we have some issues with how routing is working and how the header bars are appearing.
An empty ionic header appeared over top of the new React Native Paper header, so I brought the React Native header to the forefront using z-index. But this caused the other pages to be broken.
The solution here is that every screen still using Ionic that has a header needs to define its own <ion-nav-bar>.

This will allow the new screens to coexist with the old ones!