TimelineScrollList: more refactoring/simplifying
We don't need to keep track of so many state variables independently for what time is loaded and what's selected and what's not. In fact, all we really need is the piplineRange, and to know what trips have been loaded. Instead of oldestLoaded etc, we'll just have loadedRange, which is memoized and depends on the values of allTrips and pipelineRange.

Instead of watching oldestSelected and latestSelected for changes and loading data that way, I refactored loadMoreTrips into loadAnotherWeek and loadSpecificWeek. These call fetchTripsInRange and then call handleFetchedTrips, which sets allTrips.
So no need for things like oldestSelected and latestSelected.