fix Multilabel update on label
The previous implementation used $broadcast to notify the infscroll list that an update needed to occur. React doesn't work like this.

Instead, we're going to use a Context Provider, which basically provides some value throughout a component tree. In this case, we're going to provide some functions that need to be called from MultilabelButtonGroup so that TimelineScrollList can update its state.

All of the logic for resizing cards is no longer needed; FlashList is smart enough to do this on its own.