reorganize webpack config to allow build
As long as I've been working on the migration, I've only been using the serve configuration and hadn't figured out how to build the app.
Since we're using webpack, we should use the 'production' mode, which further optimizes your bundles. But this makes things a little trickier.

We don't use explicit annotations in our AngularJS code and this causes it to break upon minification
https://docs.angularjs.org/error/$injector/strictdi

Thus, the prod configuration of webpack requires a little extra work to get past that.