take explicit 'name' arg in angularize()
I had tried to make angularize() as simple and concise as possible. In doing so, I had it grab the name of the component using .name (ie the name of the function itself) to define the Angular component, instead of having to specifically write it out.
I thought that was a neat trick. Unfortunately, Webpack spoiled my fun; the minification process randomizes the names of the components. So we'll have to type it out and pass it as another parameter. Argh!

Fortunately, by now many of the components, at least that I've written, don't even need to be angularized anymore. Only the top-level, entry points need to be angularized, so anything that's only included in a React component can just be imported directly.