fix leaflet attribution covering map
By default, .leaflet-pane has a z-index of 400, and thus covers up the control when it's in an absolutely positioned context. This ensures the attribution control will show on top.

Follow-up to this:
@shankari My understanding (from what I looked up) is that displaying attribution to OpenStreetMap is required, while the attribution to Leaflet is optional.
Maybe we can remove the "Leaflet" part from the embedded maps in the Label screen?

Then maybe on the larger maps, where there is space for it, we can include both in the attribution.