implement React Navigation
This further breaks extracts parts of LabelTab into LabelScreen.
This will allow LabelTab to serve as the root and include a Navigator, which can route to either LabelScreen or LabelDetailsScreen (in progress).

Anything from LabelTab that is used by either LabelScreen or LabelDetailsScreen will have to be exposed in the LabelTabContext.
It is a little annoying to share all these values between components, but it's better than shoving all the logic in one place.

Clicking a trip card does navigate to LabelDetailsScreen, so the routing is working at this point. All that's left to do is rewrite the details page.