fix ionic viewable area height
This is an issue with ionic, where the content view has actually extended partway into the tabs so that the bottom of the viewable area has been covered up.
This is a temporary override to fix it. Soon the tabs themselves will be in React Native and this will be irrelevant.