wait for $ionicPlatform BEFORE DynamicConfig
I tried to make this a little faster by using Promise.all. It turns out that DynamicConfig actually depends on ionicPlatform.ready() being already done, so they can't overlap in execution after all.
Also including Logger to diary.js. It was formerly included in infinite_scroll_list.js but I removed that, so it needs to be included here.