join page: fix 'welcome' translation
When we swapped out ng-translate for ng-i18next, this text in particular stopped translating correctly.
I guess it's a slight difference in how the two directives work.

With the emoji inside the element alongside the translation key, I think i18next is actually looking for a translation key that includes the emoji, which does not exist.
To remedy this, let's just use the `| i18next` filter, which allows us to interpolate the translation without the emoji, and then just include the emoji after.