update LabelDetailsScreen
The details screen will now show percentages for the sensed mode of travel.
If the trip has more than one section, it will list the sections displaying their duration and mode of travel.

I also adjusted margins and spacing throughout the details screen. Refactoring this later to clean it up would be a good idea.