add 'mode percents' indicator to TripCard
Floating at the top of the map, we will show an indicator that displays the inferred (or sensed) modes of travel and how much of the trip they were used for.
Since this is specific to the TripCard, I didn't see a need to make a separate component for this. If we start needing it other places, we can consider it then.