fix OSM tiles not loading in built application
In the Content Security Policy, img-src was set to accept 'self' and 'ionic: data: https://*.tile.openstreetmap.org'. I believe this was necessary for the purpose of using ui-leaflet.

Now, we have LeafletView, which is independent of Ionic and AngularJS. It retrieves tiles directly from https://tile.openstreetmap.org.
Updating the policy as such resolves the issue.