don't detach main view on navigate to LabelDetails
The `detachPreviousScreen` option is true by default and means that when you navigate to a new screen, the old one is detached so it can be removed from memory.
In the case of the label details screen, we are going to go right back to the main label screen from the details page. So we shouldn't detach it.

This also fixes the issue of the scroll position resetting when going back from the details page.