add chartJS to project
ChartJS is a modular charting library that only includes basic functionality in the core package.
So there are a couple others we will need too:
 - chartjs-adapter-moment is so that time-based axes can be labeled using time formats resolved from moment.js. Then all we have to do is pass in timestamps and chartJS will figure out how to show them on the axes.
Fortunately, we already have moment.js so this doesn't incur much package size cost.

 - chartjs-plugin-annotation is how we will display dashed lines on the graphs (to show whether you're meeting a 'target' or goal)
Basically, it just enables drawing extra information on top of the chart.