implement BarChart component using ChartJS
Our new `BarChart` component is a wrapper around ChartJS's `Bar`. This should make it convenient to draw either vertical or horizontal bar charts, and easily show dashed lines for any number of 'target' values.

This is now demonstrated in the 'Chart' tab of the dashboard, where horizontal bar charts will show for Distance, Trips, Duration, Speed.

metrics.js was tweaked so that the data is formatted for the new BarChart component (and with timestamps on y axis and values on x axis)