use 'queriedRange' instead of 'loadedRange'
There was a logic error in this implementation of the label screen with how it handles subsequent fetches.

loadedRange was defined as the range between the start_ts of the earliest loaded trip and the end_ts of the latest loaded trip. I realized this is not actually what we should be tracking because we base subsequent queries off of it, and that's problematic.

If you tried to load the previous week, but there were no trips during that week, you would get stuck and not be able to go past the empty week. On every request, it would find no trips in the previous week, and so loadedRange would never be updated.

We can't use the loaded trips as a way of determining what has been fetched; we do actually have to keep track of what was requested.
This will be called `queriedRange`. It is what will show on the datepicker too. So then, there's really no use for `loadedRange`.