fix autoreload happening twice
This is just a fix for an annoying issue affecting the dev environment.
Ever since adding Webpack, whenever a JS file is changed and saved, the autoreload could sometimes happen twice. This is because autoreload (which uses chokidar) watches the entire 'www' directory for changes.
It observes one change (in www/js/) when the JS file is saved, then another change (in www/dist/) when webpack is done bundling.

A workaround this problem is to have chokidar ignore the 'www/js/' folder. This way if a JS file is changed, it won't autoreload until webpack is done. Meanwhile if an HTML file is changed, it will autoreload right away as it doesn't need to wait for webpack.

The loose assumptions here are (1) every file in www/js/ is being bundled (which is true unless unused files are left around) and (2) no files inside www/ but outside www/js/ ARE being bundled.

The second assumption is probably not perfectly true; some CSS files are bundled and so these might still cause a reload to happen twice, but this is at least better than before