package.cordovabuild.json: add "build-dev" script
This configuration could be used to build the app using the 'dev' webpack config (that is, without minifying the Javascript).

Note that there is a significant performance difference between 'dev' and 'prod'. The Label screen works much more smoothly in production (likely because FlashList is optimized for 'production' https://shopify.github.io/flash-list/docs/performance-troubleshooting)
So we might use this build for testing, but I don't think we should ever deploy a package generated by 'build-dev'.