charts: use chartjs-adapter-luxon to fix timezones
To represent the dates on the charts, we would like to use unix epoch timestamps, which represent the start of the day in UTC.
However, ChartJS by default uses the user's local time when it parses timestamps. We would like to override this behavior and use UTC, but the Moment.js adapter for ChartJS doesn't support specifying time zones.

So, this is the point at which we need to add Luxon to the project. Then we can use the Luxon adapter for ChartJS, which does support time zones.

We should really be switching from Moment.js to Luxon anyway. We still rely on Moment.js in many other places in the app, so we have to keep both around for now.
Moment.js is very heavy; fortunately Luxon is more lightweight. Before too long, we should be able to remove Moment.js altogether and enjoy a smaller bundle size :)