rework DateSelect to use React Native Paper Dates
Before now, DateSelect was using the HTML <input type='date'> element. This launched a native datepicker for iOS and Android respectively, but with some limitations such as not being able to constrain the range of selectable dates.
React Native Paper Dates offers fully featured date pickers and time pickers that integrate with the other RN Paper components we use. This library will be used in several other places of the app too.

The new DateSelect implements this and replaces the need for any HTML.

A new translation "Today" was added in en.json.
Translations for the datepicker itself needed to be registered in i18nextinit.ts.

marginLeft auto was added to the refresh button.
DateSelect had marginRight auto before, pushing the refresh button over. But It's more logical for the refresh button to control its own position using marginLeft; the result is the same either way