fix DateSelect error if clicked before ready
The initial state for selDate is what Date the DatePicker is set to. In the case of '-', a non-datelike string, we would trigger an error.
So if the DatePicker was clicked before the pipeline range was determined, we could face problems.

Using null as an initial value works fine; it will just use today's date at first and then once the pipeline range is loaded, anything outside the range is greyed out.