fix Leaflet maps getting out-of-sync
When loading in more trips, or when scrolling quickly through the label screen, sometimes the wrong map would show for a trip.
I determined this was due to the way FlashList recycles its list items and their children. https://shopify.github.io/flash-list/docs/recycling
It behaves this way in effort to maximize performance, but it requires that we take some extra steps.
Because the Leaflet maps are attached to an DOM element when they are created, we cannot expect them to re-render on their own if the DOM element becomes associated with a different trip.
To remedy this, we'll keep track of the geojson ID and see if it changes between renders. If so, clear out the map and repeat the initialization logic from before.