LabelTab refactor
The variables that track state in the label tab have been significantly refactored.
The previous flow was:
--receive composite trips -> allTrips -> filtering -> displayTrips -> unpack composite trips -> listEntries.

Now, we will do:
--receive composite trips & unpack them -> timelineMap -> filtering -> displayedEntries

The biggest difference is that unpacking composite trips (ie extracting the start confirmed / end confirmed place, if necessary) happens right after the trips are read in.
This simplifies the LabelTab because now there are essentially only two main timeline data structures to keep track of: 1) a complete, unfiltered list, and 2) the filtered list which will be displayed.

For (1), timelineMap, I decided to use the Map object. It is a simple key-value data structure, but unlike an object, it preserves order, which is an important consideration for our timeline.
The benefit of it being keyed is 1) no possibility of duplicates and 2) easy lookup.
-- for the Map object to be used, I needed to enable a TypeScript setting that would allow the use of Map in ES5.

The second structure, displayedEntries, is much like before. It automatically derives its state from a) timelineMap and b) the filter that is active, and it represents the actual list of timeline entries that are shown in the list.

Also in this commit, some functions were extracted into timelineHelper.ts, which is gradually replacing the factory in diary/services.js, and other functions were extracted to the React hook useImperialConfig, which will replace config/imperial.js.
Until the migration is further along, I need to keep both the old and new files around.