settings tables migration to React - Abby's first migration task (#992)
* switched collect and sync lists over to react

* inserted a placeholding div control-list-item

This is not yet complete, but is an idea of where the schedule table might eventually be

* patch from Jack

incorporating a patch from Jack with a solution to some of the issues with scheduled notifications, as well as cleaning up some of the style around the control-data-table (s) I converted earlier

* added keys to rows

found an error in the console alerting to the fact that each row should have a unique key, so I added keys
one concern: the upcoming notifications keys end up being the dates - could those ever be repetitive?

* removal of non-relevant code

took out the code that was a part of a different effort to prevent confusion, also removed commented out code form replacing elements

* replacing hard coded strings

the "Upcoming Notifications" and "Dummy Notification in 5 Seconds statements were hard coded, I've added them as internationalized strings and updated the en.json file

* start at storing data in better place

I started to convert the data storage, data is fine in general-settings.js but is failing to make it through main-control.html to ControlDataTable.jsx

* spelling

* starting to work with other rows of profile

this is where I'm struggling with the parameterization of i18n keys

* setting up issue in using key as parameter

Added in code so the broken setting row is enabled, note the console.log statement demoing the different ways the info comes through

* fixing strings

in Angular, string parameters need to be created as "'string'" - this should resolve some of the issues that I have been experiencing here

Co-authored-by: Jack Greenlee <JackAGreenlee@gmail.com>

* text working on sample list item

after I learned how to pass strings, I was able to render the row as a list item, which will hopefully eventually be good for showing everything as a list, and can have icons built in

Icons and passing through functionality are next

* carry data storage change through

switching from plugin storage to scheduler storage

Co-authored-by: Jack Greenlee <JackAGreenlee@gmail.com>

* polishing up - removing print statements, altering log

after fix from Jack (2nd place to update where notifs come from) added in optional chain to reduce errors in table, took out diagnostic log statements, updated Logger.log to show num, time, and first notif

* update comments, eliminate changes not a part of this PR

the setting-row process has been separated to More profile migrations #994

* action parameters for setting rows - some work

have a rough outline going of some setting rows that say and do the right thing! More complex actions, some text, and icons still need lots of work

* comments and spacing

I added a bunch of comments and whitespace to make it easier for myself to see what needs to be done -- I'll reformat once I have things working

* more row migrations

most rows now migrated, still need to handle some with slightly unique functionality, styling with icons also needed!

* rough draft - two rows unconverted

all rows ugly but functional but need guidance on reminder time and demographics button

* fixing strings - i18n

* enabling icons - copy Jack's related commit

https://github.com/e-mission/e-mission-phone/pull/993/commits/6578d8ab71e1e107869dc5a5e2ef389686099d1c

* got icons in

still missing some visual cues (toggle switches, expansion flippers)

* tidy code - still ways to go

* combine logout and copy (eliminate copy)

* convert to use of IconButtons

now the button is what has action, not the entire row - there's a bug requiring an empty onPress in the list Item for the one in IconButton to work see issue https://github.com/callstack/react-native-paper/issues/3898 and hopefully soon to be fix  https://github.com/callstack/react-native-paper/pull/3908

* introduce toggle switches

the two items that had switches are now interactive again, there is work to be done with getting the switches to have the proper starting value

toggle switch: https://callstack.github.io/react-native-paper/docs/components/Switch/

* experiment with list accordian

converted the user data section to a list accordian - works great! Unfortunately relies right now on a lot of hard-coding

* use of .map for creating the collapsing list!

first-run at creating a drop-down part of the menu with map and passing in a data structure

* both drop-downs converted to ExpandMenu components

fully converted the userData and devZone to the react element, creates an accordian element that has SettingRow and ControlDataTable elements

* profile settings component draft

* profile settings in general settings

* first set settings now in ProfileSettingsComponent

creating an over-arching component for the ProfileSettings -- still passing in functions, some things have bugs/functional differences noted in comments

* different method of passing info to Profile Settings

* introduce Demographics Setting Row

* rough sketch everything in ProfileSettings

working towards full migration and pulling everything into a single react Component. This is not working perfectly, there are a lot of things to handle that are currently commented out

* more of migration to ProfileSettings component

removing some commented out code
in ExpandMenu.jsx, props is now the listed parameter, this allows to get the default parameter props.children to populate the inner components. the section Title is accessed with props.sectionTitle

* adding a value in the scope for the accuracy setting

working on the switches, and trying to get their displayed value to match the value of the think they control

* adding additional argument to angularize call

this is needed from changes in #993

* cleaned up irrelivant codes and comments

* Revert "fixing merge conflicts"

This reverts commit ffd2bc2bf0b1595566b7287f64d57d9b91a8ad3b, reversing
changes made to a3097ac7f383ef408e326c1c3d5f8c2ef1e7f05f.

---------

Co-authored-by: Jack Greenlee <JackAGreenlee@gmail.com>