make labeled trips stay 30s before disappearing
The expected behavior, prior to any of the migration, was that after labeling trips, there is a 30 second delay before they disappear from "To Label" and get filtered out.
This reinstates that behavior by use of a flag called 'justRepopulated'. Any trip that has  'justRepopulated' will be immune to filtering.

Getting setTimeout to work in React was not very straightforward. When trying to access state from inside a setTimeout, we end up with what's called a 'stale' closure. (some comments left in-code: see https://legacy.reactjs.org/docs/hooks-faq.html#why-am-i-seeing-stale-props-or-state-inside-my-function)

This means we end up getting the value of state at the time the setTimeout was defined, not at the time it is executed. The only way I found around this is keeping track of a ref that mirrors the relevant state, and then access the ref's `current` property.