fix updating trips from details screen
When a trip is labeled from the details screen, the view should update accordingly. But because the trip object was passed as a routing parameter, reactivity was broken and the labels did not show until the user navigated back to the main label list screen.

Instead, we should pass the trip ID and have the details page get the trip object from the timelineMap of the LabelTabContext. This way, when a trip is updated inside timelineMap, its changes will be received downstream through useContext.