🗃️ Store the dynamic config as local storage in addition to RW document
So it is stored both in the webview local storage and the SQLite database.

While reading, we might only have the RW document in the older versions, and we
use it to populate the local storage as needed for backwards compatibility and
a gradual migration.

We might want to remove the RW document in the future, but maybe at that time
we will have more reliable SQLite storage anyway.