ProfileSettings: track collectSettings w/ useState
In general-settings, the `$scope.settings` object is used to store all of the current values of options that are represeted by the Profile Tab.
As we migrate to ProfileSettings, this should be stored as state in React. This change handles only `collect` settings; everything that was previously stored under `$scope.settings.collect`.

We accomplish this with useState and a new function `refreshCollectSettings`, which is called 1) on config load and 2) whenever any setting relating to collection is changed

`settingsScope` needed to be retrieved in a different way. Previously it was passed from AngularJS a prop; this caused the ProfileSettings component to be re-initialized every time any scope value changed. This made it impossible to persist state in ProfileSettings.
Now we will retrieve the scope with a DOM lookup; we find the root element of the scope, then ask Angular for the scope attached to that element.