diaryHelper: parse dates with respect to timezone
The app interface should be showing the date and time of a trip/place *based on the timezone that the trip was recorded in.*
So when we want to parse a trip and get human-readable dates, we can't just use the epoch timestamp. Epoch timestamps only tell us what the date was at a given moment in UTC. We need to consider the timezone of the trip.
Thus, all these function need to be modified to accept the `fmt_time` strings (which are ISO 8601 and include timezone +/-). Then we can employ moment.parseZone and format.

Result -- dates are now calculated and displayed properly: with respect to the timezone and not abnormal if the date in UTC differed from the actual date.

note: the trip object's `display_duration` was removed because it is not even used anywhere in the codebase; `display_time` is consistently used instead.