update CSP to fix enketo checkboxes' inline SVGs
Enketo uses Sass (.scss) to define its styles. The way that the checkbox 'checks' are defined by Enketo is by defining an inline SVG as the background image to the checkbox.
See inline SVG usage: https://github.com/enketo/enketo-core/blob/166af8c0d9bfe37ad65c4dbc5a805b458bae0948/src/sass/formhub/_mixins.scss#L16

This does not bundle properly in production builds, causing the checkmark to not appear. This was difficult to troubleshoot because it only happens in production. But on further investigation, I found the cause is the Content Security Policy, which is enforced more strictly in production. Updating this to allow inline evals fixed the issue.