support iOS<14 by transpiling dependencies in prod
In order to support iPhone 6, we must support iOS 12. Unfortunately Safari did not implement static properties in classes until version 14, and many of our dependencies use this modern syntax in the JS they expose.

To remedy this, we must tell babel to perform extra transpilation for production builds. This will lengthen the build process by about 1 minute, but it should guarantee that the JS exposed by the bundle is safe.

Awaiting testing