refactor DateSelect+FilterSelect into NavBarButton
NavBarButton is new and will contain the styling + presentational components that are common to both DateSelect and FilterSelect. It could be used in other tabs too if we wish to have more buttons embedded in the top navbar.

So then, DateSelect and FilterSelect will both be composed of a NavBarButton with different contents inside the button.

FilterSelect has reworked to no longer use HTML.
Instead of an HTML <select>, whose behavior is dependent on OS/browser implementation, it will launch a Modal with radio buttons. This implementation is consistent across platforms and more stable.
In the process, this change also fixes issues in which the text of the <select> was not updating.