allow displayedEntries to be null before load
There's an important difference between 'no travel to show because we are still loading' and 'no travel to show because the request returned no trips'.
So before anything is loading, instead of initializing to an empty array, displayedEntries should initialize to null.
If timelineMap is null, displayedEntries will be null too.

Then, the "No travel to show" banner will only show if displayedEntries does exists, but has a length of 0.