update QRCode sharing methods
In migration, the way QR codes are displayed changed. In order to download, the React-generated svg needed to somehow be converted to another format. This is demoed in the react-qr-code here: https://rosskhanas.github.io/react-qr-code/ with code here: https://github.com/rosskhanas/react-qr-code/blob/ade96f86a85d8018e94663462a18e609029dacaa/demo/src/components/App.js#L30

This method finds the svg displayed, and encodes it as an Image object, draws that Image on a "canvas" element, and converts the canvas to a png that is able to be shared/downloaded. The `querySelector` is particular to the result of angularizing the react component