refactor fetching notifs to ensure scheduled first
Particularly on initial startup, the notifications were not done scheduling before they were fetched and so they were often undefined or partial.

When the `isScheduling` var is true, the get method now awaits resolution of the getScheduled promise before fetching. The fetch itself is awaited in the notificationsRefresh of ProfileSettings so that notifications are present before displaying them

-> hoping this can be simplified upon full migration into React