fix duplicate activities
The duplicated activities were caused by the fact that the entire updated list of activities was being pushed onto the existing list of activities. By zeroing out the existing list (`timelineEntryField` which I noted was empty on primary loading) the updated list (with all existing activities and the new one) is pushed into the list. - this also explains why there was no error for the first activity added, as there were not existing activities to re-push