i18n: only use first locale if multiple detected
On some systems (apparently Android or at least Pixel), the `locales` array may include locales from multiple languages in order of preference. In this case, we just want to use the first language that we have translations for, and then skip the rest.

To fix this let's switch to `for .. of` syntax, and then we can `break` out of the loop once we find a supported language. (the `foreach` syntax does not have a clean way to exit the loop like this)