DateSelect: default to "Today"
if pipelineRange is not defined, this means 1) we are still initializing or 2) the server has no processed time series entries for this user
In this case, we are not going to have anything to show. Instead of " - / - " which was shown on the datepicker before, let's show Today.
In this case, we also want min and max dates to be constrained to the current date - this way the user does not go looking for days that don't exist.