TimelineScrollList: adjust noTravelBanner
if pipelineRange is undefined, the 'no travel' banner should always show (this is what the user will see when they first log in.
Otherwise, the 'no travel' banner should also show if there are no trips in view. However, we still want a loading spinner to show as long as the pipelineRange is defined.
To accomplish this, let's extract the banner to a new const called 'noTravelBanner' and reorder the if-else chain according to the expected behavior.

This commit also internationalizes the 'no-travel' and 'no-travel-hint' text.