unify leaflet div icons CSS into LeafletView
Some of the styling for the leaflet div icons was scattered between main.diary.css and LeafletView.jsx after unifying the icons themselves.
Let's unify and have all the relevant styles kept in LeafletView.jsx.

In the process of this, we can replace all references to CSS variables; let's instead use colors from the color palette we get from `useTheme`.