fix leaflet styles not applying correct flavor
In plain HTML, <style> tags work globally. This means a style declared in a child can affect its parents, siblings, and even distant cousins. This is problematic now that we are programatically setting the colors.

For draft trips, because they have a 'flavor' different than the main app theme, their <style> tags contain different colors than non-draft trips.

But the draft trip <style> tags were still being overridden by the other <style> tags because they are all global.

To remedy this we must prefix those CSS selectors with an ID specific to the current map.
This way the styles are scoped to the map currently being defined.