change screen backgrounds to bluish gray
the other gray background clashed a bit with the blue-tinted cards. This gray is a nearby color to the blue scheme already in use

For the label screen it is controlled appTheme and applied in LabelListScreen. Since the profile and dashboard are not converted, the same color is applied to the "pane" class in style.css