fix intro skipping to demographics survey
While completing the onboarding / 'intro', the Save OPCode slide would be skipped. The next slide, which is the demographics survey, would load after ~1 second and then overtake the Save OPCode slide.

I am guessing that when you move to the next slide of an ion-slide, it looks ahead to the next slide and creates it ahead of time.

By systematically commenting out code, I isolated the issue and determined it only happens when the Enketo Form object is `init()` (service.js -> _loadForm).
I think there is some code there that focuses the form and ensures the form is visible. So with it being off-screen at the time it is initialized we get strange results.

So the workaround here is to prevent the survey from being initialized before that slide is active. intro has a slideIndex we can look at, and we will only start rendering the contents of the survey slide once slideIndex reaches 4 (the fifth slide / the survey).

I also found that in EnketoDemographicsInlineCtrl, the 10-second $timeout that waits for the slide to be in place is too low. 50 is safer.

Overall, this is a pretty temporary solution; the intro will be rewritted in React before too long.
We will use something other than ion-slide, and very likely redesign the onboarding process in general.