fix 'no travel' showing before pipeline range set
The "No Travel" banner is supposed to show when there are no trips. However, it was showing for a couple seconds, after app launch, but before the pipelineRange is fetched.
This is a place where we really want the spinner to show.

In LabelTab, if nothing had been processed for a user yet, getUnprocessedLabels would come back with a pipelineRange of {start_ts: null and end_ts: null}, but we would ignore this result.
We should setPipelineRange to this even if it's empty so that we can indiate it's already been fetched.

Then in TimelineScrollList, instead of checking if pipelineRange isn't defined, we should instead check if it *has* been set, but the end_ts is null. In this case we show the banner.
This way, if pipelineRange itself is not set yet, we get the spinner.

Also added comments to better understand the conditional logic for what's shown in the timeline scroll list.